+++ src/fmalloc.cpp
@@ -4,4 +4,5 @@
 #include "fmalloc_vector.inc"
 #include "fmalloc_altrep.inc"
 #include "fmalloc_ops.inc"
+#include "fmalloc_tensor.inc"
 #include "fmalloc_api.inc"